refactor: complete server transport redesign cutover
What changed and why:
- Replace old transport interfaces with new v2 implementations
- Remove artificial lifecycle ceremony (open/close/is_open)
- Eliminate transport state checks from coordinator
- Enable foundation for multi-client architecture
- Drop _v2 suffix from session, coordinator, and transport files

The original transport design artificially constrained servers to
lifecycle management that added complexity without value. The new
design focuses purely on message routing with explicit connection
management, creating symmetric client/server patterns.

Impact:
- Cleaner abstractions with single responsibilities
- Eliminated artificial state management complexity
- Foundation ready for multi-server client architecture
- All existing tests pass with no API changes
- Transport layer now handles pure message routing
